Sign Up and Logo Animation is ready
Sign Up page has been done, we need to make sure the Sign up and the Sign in page are identical.

Logos have been downloaded.

Loading animation when webpage being opened, is prepared and ready to be used.
IMPORTANT: This animation can not be used as a loading overlay between pages.